requirements.txt
assertpy==1.1
mysql_connector_repackaged==0.3.1
pytest==7.1.2
requests==2.22.0
